requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
aiofiles==23.2.1
asyncio-mqtt==0.16.1

